add save/update membrane gaze data
